+++ src/briefy/leica/config.py
@@ -61,3 +61,8 @@ SCHEDULE_DAYS_LIMIT = config('SCHEDULE_DAYS_LIMIT', default=1)
 
 # Cache
 ENABLE_CACHE = config('ENABLE_CACHE', casts.Boolean(), default=False)
+
+# default 48 hs
+LATE_SUBMISSION_SECONDS = config('LATE_SUBMISSION_SECONDS', default='172800')
+# default 24 hs
+BEFORE_SHOOTING_SECONDS = config('LATE_SUBMISSION_SECONDS', default='86400')